refactor: BusNotice 공통컴포넌트화
