+++ src/pages/Bus/BusCoursePage/BusCoursePage.module.scss
@@ -53,65 +53,3 @@
   font-size: 18px;
   line-height: 160%;
 }
-
-.notion-info {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 16px 24px;
-  border-radius: 8px;
-  background: #f1f8ff;
-  box-shadow: 0 1px 1px 0 #00000005, 0 2px 4px 0 #0000000a;
-  margin-top: 24px;
-  margin-bottom: 32px;
-
-  &__info-title {
-    overflow: hidden;
-    color: #175c8e;
-    text-align: center;
-    text-overflow: ellipsis;
-    font-family: Pretendard, sans-serif;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 160%;
-  }
-
-  &__placeholder {
-    min-height: 89.6px;
-  }
-
-  @include media.media-breakpoint(mobile) {
-    justify-content: center;
-    font-size: 14px;
-    font-weight: 500;
-    padding: 12px 16px;
-    margin-top: 0;
-    margin-bottom: 8px;
-  }
-}
-
-.notion-info-closed {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 16px 24px;
-  margin-top: 24px;
-  margin-bottom: 32px;
-  background-color: #fff;
-  color: #fff;
-
-  &__button {
-    height: 35px;
-  }
-
-  @include media.media-breakpoint(mobile) {
-    justify-content: center;
-    font-size: 14px;
-    font-weight: 500;
-    padding: 12px 16px;
-    margin-top: 0;
-    margin-bottom: 8px;
-    background-color: #fff;
-  }
-}
