+++ src/pages/Bus/BusCoursePage/BusTimetable/BusTimetable.module.scss
@@ -2,7 +2,7 @@
 
 .template {
   width: 795px;
-  margin: 64px auto 0;
+  margin: 23px auto 0;
 
   @include media.media-breakpoint(mobile) {
     min-height: 350px;
